Relations are now properly removed when one side of the relation is destroyed.
This necessitated adding a new event, 'destroying'. It's trigger right before
an object is destroyed, when its state is still available. This is needed in
order to get access to the list of relations to be removed.

The 'destroyed' event stays as it was - the object is already gone when it
triggers.